requirements.txt
customtkinter
pylix
keyring
argon2-cffi
cryptography